// src/interfaces/TicketData.ts
export interface TicketData {
	id: number;
	clientApplicationId: number;
	eventId: number;
	section: string;
	sectionId: number;
	sectionMapName: string;
	sectionType: number;
	row: string;
	seat: string;
	seatFromInternal: string;
	hasSeatDetails: boolean;
	hasSeatDetailsUS: boolean;
	availableTickets: number;
	listingPreviewPriceAndFeeDisclosure: { hasValue: boolean };
	inventoryListingScore: {
		discount: string;
		starRating: number;
		dealScore: string;
		seatQualityScore: string;
	};
	showRecentlySold: boolean;
	availableQuantities: number[];
	ticketClass: number;
	ticketClassName: string;
	maxQuantity: number;
	hasListingNotes: boolean;
	listingNotes: {
		listingNoteId: number;
		listingNoteContentId: number;
		formattedListingNoteContent: string;
		listingNoteTypeId: number;
		showToBuyer: boolean;
		hideInMock: boolean;
		siteAddedListingNote: boolean;
		aisleListingNoteWithSplit: boolean;
		listingNoteDescriptionContentId: number;
		formattedListingNoteDescription: string;
	}[];
	rowId: number;
	isUsersListing: boolean;
	isPreUploaded: boolean;
	rowContent: string;
	rawPrice: number;
	price: string;
	sellerNetProceeds: string;
	priceWithFees: string;
	ticketTypeId: number;
	ticketTypeGroupId: number;
	listingTypeId: number;
	listingCurrencyCode: string;
	buyerCurrencyCode: string;
	faceValue: number;
	faceValueCurrencyCode: string;
	vfsUrl: string;
	formattedActiveSince: string;
	calculatedBestValue: string;
	formattedDealScore: string;
	isSeatedTogether: boolean;
	sellerUserId: string;
	showVfsInListing: boolean;
	hideSeatAndRowInfo: boolean;
	sellerHideSeatInfo: boolean;
	aipHash: string;
	isMLBVerified: boolean;
	isStanding: boolean;
	createdDateTime: string;
	isHighestListingScore: boolean;
	isMostAffordable: boolean;
	isSponsored: boolean;
	isCheapestListing: boolean;
}